don't mutate donationCounter objects
This fixes a bug where you are looking at a milestone. You click the
owner to view their profile, then you click the browsers back button.

Upon viewing the milestone again, it appears that a cached version of
the milestone is returned when calling MilestoneService.subscribeOne.
This cached version has had the donations counters totalDonated &
currentBalance mutated previously & set to BigNumber instances. An error
is throw when passing the BigNumber instances to utils.fromWei.